+++ src/num/grain/mod.rs
@@ -7,7 +7,7 @@
 //
 
 mod cast; // Cast, PrimCast, PrimJoin, PrimSplit, cast!
-pub mod niche; // MaybeNiche, NonNiche*, NonZero*, NonValue*|NonExtreme*, ne!, nz!
+pub mod niche; // MaybeNiche, NonNiche*, NonZero*, Non<Max|Min|Value>*, nm!, nv!, nz!
 mod prim; // PrimFloat, PrimInt, PrimScalar, PrimSint, PrimUint
 pub mod wide; // lane!