+++ src/text/unicode/scalar/scalar/traits.rs
@@ -11,11 +11,11 @@ use crate::{MismatchedCapacity, paste, text::unicode::scalar::*, unwrap};
 macro_rules! char_core_impls {
     () => {
         char_core_impls![
-            char7 + Self(unwrap![some NonExtremeU8::new(0)]),
+            char7 + Self(unwrap![some NonMaxU8::new(0)]),
             char8 + Self(0),
             char16 + Self(unwrap![some NonSurrogateU16::new(0)]),
             charu + Self(unwrap![some NonNiche::<u32>::new(0)]),
-            charu_niche + Self(unwrap![some NonExtremeU32::new(0)])
+            charu_niche + Self(unwrap![some NonMaxU32::new(0)])
         ];
     };
     ($( $name:ident + $default:expr ),+ $(,)?) => {